project1/requirement.txt
flask==3.0.3
mysql-connector-python==9.0.0
requests==2.32.3